+++ src/commands/switch_aws_profile.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { getAWSProfiles } from '../aws_api';
+import { getAWSProfiles } from '../api/aws_api';
 
 export default function createSwitchAWSProfileCmd(context) {
   return vscode.commands.registerCommand('extension.switchAWSProfile', async () => {