+++ src/ui/ssm_key_codelens_provider.ts
@@ -1,95 +1,73 @@
 import * as vscode from 'vscode';
-import { extractSSMKeysFromFile } from '../tf_ssm_file_parser';
-import { showResource, ShowResourceResult, TerraformError } from '../tf_state_api';
-import { dirname } from 'path';
-import * as bluebird from 'bluebird';
+import { extractSSMKeysFromFile } from './tf_ssm_file_parser';
+import SSMKey from './ssm_key_model';
+import TFLocalResourceCache from './tf_local_resource_cache';
+import { TerraformError } from '../api/tf_state_api';
+import TFResource from './tf_resource_model';
 
-// note(itay): a TFResource is something like data.some_place.some_key
-// the ResourceName is all apart of some_key and KeyName is some_key
-interface TFResource {
-  ResourceName: string;
-  KeyName: string;
-}
+class SSMKeyCodeLens extends vscode.CodeLens {
 
-async function enrichSSMKeyPath(path, document, awsProfile, extensionContext) {
-  const resourceKeyParts = await bluebird.map(path.split("/"), async resourceInPath => {
-    if (resourceInPath.indexOf("$") > -1) {
-      const leftBracketIndex = resourceInPath.indexOf("{");
-      const rightBracketIndex = resourceInPath.indexOf("}");
-      const resourceString = resourceInPath.slice(leftBracketIndex + 1, rightBracketIndex);
-      const resource = getResourceFromResourceString(resourceString);
-      const resourceKeyValue = await getAndSetValueFromResource(resource, document, awsProfile, extensionContext);
-      return resourceKeyValue;
-    } else {
-      return resourceInPath;
-    }
-  }, { concurrency: 1 });
-  return resourceKeyParts.join("/")
-}
-
-async function getAndSetValueFromResource(resource: TFResource, document, awsProfile, extensionContext) {
-  const storeInKey = `${resource.ResourceName}.${resource.KeyName}`;
-  let storeValue = extensionContext.globalState.get(storeInKey);
-  if (!storeValue) {
-    const resourceObj: ShowResourceResult = await showResource(resource.ResourceName, dirname(document.fileName), awsProfile) as ShowResourceResult;
-    if (resourceObj.Error) {
-      if (resourceObj.Error === TerraformError.NoTerraformInstalled) {
-        vscode.window.showErrorMessage(`The extension vscode-terraform-aws-ssm must have terraform CLI installed to work properly`);
-      } else if (resourceObj.Error === TerraformError.FailedToLoadBackend) {
-        vscode.window.showErrorMessage(`You must run terraform init on the directory in order for Edit SSM Key to work`);
-      } else if (resourceObj.Error === TerraformError.EmptyResponse) {
-        vscode.window.showErrorMessage(`terraform state show ${resource.ResourceName} returned an empty response`);
-      } else {
-        vscode.window.showErrorMessage(`Current profile ${awsProfile} has no permissions to load terraform state.`);
-      }
-      throw new Error("Stop SSMCodeLensProvider parsing step due to error");
-    }
-    storeValue = resourceObj.Properties.get(resource.KeyName);
-    extensionContext.globalState.update(storeInKey, storeValue);
+  constructor(
+    private ssmKey: SSMKey,
+    range: vscode.Range,
+    command?: vscode.Command
+  ) {
+    super(range, command);
   }
-  return storeValue;
+
+  async createCommand(cache: TFLocalResourceCache, awsProfile: string): Promise<vscode.Command> {
+    const ssmKeyPath = await this.ssmKey.getPath(cache, awsProfile);
+    if (ssmKeyPath.Error) {
+      this.handleResourceError(ssmKeyPath.Resource, ssmKeyPath.Error, awsProfile);
+      return Promise.reject(ssmKeyPath.Error);
+    }
+
+    return Promise.resolve({
+      command: 'extension.editSSMKey',
+      title: 'Edit SSM Key',
+      arguments: [ssmKeyPath.Path]
+    } as vscode.Command);
+  }
+
+  handleResourceError(resource: TFResource, error: TerraformError, awsProfile: string) {
+    if (error === TerraformError.NoTerraformInstalled) {
+      vscode.window.showErrorMessage(`The extension vscode-terraform-aws-ssm must have terraform CLI installed to work properly`);
+    } else if (error === TerraformError.FailedToLoadBackend) {
+      vscode.window.showErrorMessage(`You must run terraform init on the directory in order for Edit SSM Key to work`);
+    } else if (error === TerraformError.EmptyResponse) {
+      vscode.window.showErrorMessage(`terraform state show ${resource.ResourceName} returned an empty response`);
+    } else if (error === TerraformError.UnsupportedResourceType) {
+      vscode.window.showErrorMessage(`Unsupported ResourceType '${resource.ResourceType}'. Only 'data' resource is supported.`);
+    } else {
+      vscode.window.showErrorMessage(`Current profile ${awsProfile} has no permissions to load terraform state.`);
+    }
+  }
 }
 
-function getResourceFromResourceString(resourceString) {
-  let resourceParts = resourceString.split(".");
-  //note(itay): We don't need the last part as it's a key from the resource
-  const key = resourceParts.pop();
-  return { ResourceName: resourceParts.join("."), KeyName: key };
-}
-
-export default class SSMKeyCodeLensProvider implements vscode.CodeLensProvider, vscode.Disposable {
-  private codeLensRegistrationHandle?: vscode.Disposable | null;
+export default class SSMKeyCodeLensProvider implements vscode.CodeLensProvider {
   private extensionContext: vscode.ExtensionContext;
 
   constructor(_extensionContext) {
     this.extensionContext = _extensionContext;
   }
 
-  dispose() {
-    if (this.codeLensRegistrationHandle) {
-      this.codeLensRegistrationHandle.dispose();
-      this.codeLensRegistrationHandle = null;
-    }
+  async provideCodeLenses(document: vscode.TextDocument, _token: vscode.CancellationToken): Promise<vscode.CodeLens[] | null> {
+    const ssmKeys = extractSSMKeysFromFile(document.fileName, document.getText());
+    return ssmKeys.map((ssmKey: SSMKey) => {
+      const start = new vscode.Position(ssmKey.Line, ssmKey.Column);
+      const end = new vscode.Position(ssmKey.Line, ssmKey.Offset);
+      return new SSMKeyCodeLens(ssmKey, new vscode.Range(start, end));
+    });
   }
 
-  async provideCodeLenses(document: vscode.TextDocument, _token: vscode.CancellationToken): Promise<vscode.CodeLens[] | null> {
-    let items: vscode.CodeLens[] = [];
-    const AWSProfile = this.extensionContext.globalState.get("AWSProfile") as string;
-    const ssmKeys = extractSSMKeysFromFile(document.getText());
-    await bluebird.map(ssmKeys, async key => {
-      const enrichSSMKeyValue = await enrichSSMKeyPath(key.Path, document, AWSProfile, this.extensionContext);
-      let editSSMKey: vscode.Command = {
-        command: 'extension.editSSMKey',
-        title: 'Edit SSM Key',
-        arguments: [enrichSSMKeyValue]
-      };
-      const start = new vscode.Position(key.Line, key.Column);
-      const end = new vscode.Position(key.Line, key.Offset);
-      items.push(
-        new vscode.CodeLens(new vscode.Range(start, end), editSSMKey),
-      );
-    }, { concurrency: 1 });
-
-    return items;
+  resolveCodeLens?(codeLens: vscode.CodeLens, token: vscode.CancellationToken): vscode.ProviderResult<vscode.CodeLens> {
+    const awsProfile = this.extensionContext.globalState.get("AWSProfile") as string;
+    let ssmKeyCodeLens = codeLens as SSMKeyCodeLens;
+    const localResourceCache = new TFLocalResourceCache(this.extensionContext.globalState);
+    return ssmKeyCodeLens.createCommand(localResourceCache, awsProfile)
+      .then((command) => {
+        ssmKeyCodeLens.command = command;
+        return ssmKeyCodeLens;
+      });
   }
 }